added validation in the api

+++ task-api/controllers/deleteController.js
@@ -4,11 +4,24 @@ const FormatResponse = require('response-format');
 exports.deleteTask = async (req, res, next) => {
   const { id } = req.body;
 
-  const deleteItem = await Task.deleteOne({
-    _id: id,
-  });
+  if (id === '') {
+    return res
+      .status(200)
+      .json(FormatResponse.badRequest('Id must be passed', {}));
+  }
 
-  return res
-    .status(200)
-    .json(FormatResponse.success('deleted task succesfully', { deleteItem }));
+  try {
+    const deleteItem = await Task.deleteOne({
+      _id: id,
+    });
+
+    return res
+      .status(200)
+      .json(FormatResponse.success('deleted task succesfully', { deleteItem }));
+  } catch (error) {
+    console.error(error);
+    return res
+      .status(200)
+      .json(FormatResponse.badRequest('Something went wrong', {}));
+  }
 };
