+++ task-api/controllers/updateController.js
@@ -4,14 +4,27 @@ const FormatResponse = require('response-format');
 exports.updateTask = async (req, res, next) => {
   const { id, title, description, status } = req.body;
 
-  const deleteItem = await Task.updateOne(
-    {
-      _id: id,
-    },
-    { $set: { title, description, status } }
-  );
+  if (id === '' || title === '' || description === '' || status === '') {
+    return res
+      .status(200)
+      .json(FormatResponse.badRequest('Field not passed properply', {}));
+  }
 
-  return res
-    .status(200)
-    .json(FormatResponse.success('updated task succesfully', { deleteItem }));
+  try {
+    const deleteItem = await Task.updateOne(
+      {
+        _id: id,
+      },
+      { $set: { title, description, status } }
+    );
+
+    return res
+      .status(200)
+      .json(FormatResponse.success('updated task succesfully', { deleteItem }));
+  } catch (error) {
+    console.error(error);
+    return res
+      .status(200)
+      .json(FormatResponse.badRequest('Something went wrong', {}));
+  }
 };
